+++ src/main/java/persistence/dao/OrderDAO.java
@@ -1,11 +1,11 @@
-package dao;
+package persistence.dao;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
-import persistence.dto.StoreDTO;
+import persistence.dto.OrderDTO;
 
-public class StoreDAO {
+public class OrderDAO {
     private SqlSessionFactory sqlSessionFactory = null;
-    public StoreDAO(SqlSessionFactory sqlSessionFactory){
+    public OrderDAO(SqlSessionFactory sqlSessionFactory){
         this.sqlSessionFactory = sqlSessionFactory;
     }
 }
