+++ src/main/java/persistence/dao/UserDAO.java
@@ -1,4 +1,4 @@
-package dao;
+package persistence.dao;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import persistence.dto.UserDTO;
